```
Req runs:    5  QueryEqual<mixed><NonNullable><NonIndexed> (MemOnly, EncryptionOff):     min 102.53ms (-36.21%)           max 103.34ms (-35.80%)           med 102.86ms (-36.04%)           avg 102.86ms (-36.05%)           stddev   343us (+235.82%)
```

## ☑️ ToDos
* [x] 📝 Changelog update
* [x] 🚦 Tests (or not relevant)
* [x] C-API, if public C++ API changed.